+++ packages/article-image/src/styles/responsive.web.js
@@ -57,3 +57,20 @@ export const InsetCaptionStyle = styled(View)`
     padding-left: 0px;
   }
 `;
+
+export const FullWidthCaptionContainer = styled.div`
+  padding-left: ${spacing(2)};
+  padding-right: ${spacing(2)};
+  margin: auto;
+  text-align: center;
+
+  @media (min-width: ${breakpoints.medium}px) {
+    padding-left: 0px;
+    padding-right: 0px;
+    width: 80.8%;
+  }
+
+  @media (min-width: ${breakpoints.wide}px) {
+    width: 56.2%;
+  }
+`;