+++ packages/article-main-comment/src/styles/shared.js
@@ -45,6 +45,8 @@ const sharedStyles = {
     marginTop: spacing(1)
   },
   header: {
+    // TODO: put back after design signoff
+    marginBottom: spacing(3),
     paddingBottom: spacing(4),
     paddingTop: spacing(7)
   },