+++ storybook-native/index.js
@@ -1,4 +1,5 @@
-import "babel-polyfill";
+import "core-js";
+import "regenerator-runtime/runtime";
 import { AppRegistry } from "react-native";
 import StorybookUIRoot from "./storybook";