+++ android-app/CHANGELOG.md
@@ -3,6 +3,94 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.16.28](https://github.com/newsuk/times-components/compare/@times-components/android-app@0.16.27...@times-components/android-app@0.16.28) (2019-08-28)
+
+**Note:** Version bump only for package @times-components/android-app
+
+
+
+
+
+## [0.16.27](https://github.com/newsuk/times-components/compare/@times-components/android-app@0.16.26...@times-components/android-app@0.16.27) (2019-08-27)
+
+**Note:** Version bump only for package @times-components/android-app
+
+
+
+
+
+## [0.16.26](https://github.com/newsuk/times-components/compare/@times-components/android-app@0.16.25...@times-components/android-app@0.16.26) (2019-08-26)
+
+**Note:** Version bump only for package @times-components/android-app
+
+
+
+
+
+## [0.16.25](https://github.com/newsuk/times-components/compare/@times-components/android-app@0.16.24...@times-components/android-app@0.16.25) (2019-08-26)
+
+**Note:** Version bump only for package @times-components/android-app
+
+
+
+
+
+## [0.16.24](https://github.com/newsuk/times-components/compare/@times-components/android-app@0.16.23...@times-components/android-app@0.16.24) (2019-08-23)
+
+**Note:** Version bump only for package @times-components/android-app
+
+
+
+
+
+## [0.16.23](https://github.com/newsuk/times-components/compare/@times-components/android-app@0.16.22...@times-components/android-app@0.16.23) (2019-08-22)
+
+**Note:** Version bump only for package @times-components/android-app
+
+
+
+
+
+## [0.16.22](https://github.com/newsuk/times-components/compare/@times-components/android-app@0.16.21...@times-components/android-app@0.16.22) (2019-08-22)
+
+**Note:** Version bump only for package @times-components/android-app
+
+
+
+
+
+## [0.16.21](https://github.com/newsuk/times-components/compare/@times-components/android-app@0.16.20...@times-components/android-app@0.16.21) (2019-08-21)
+
+**Note:** Version bump only for package @times-components/android-app
+
+
+
+
+
+## [0.16.20](https://github.com/newsuk/times-components/compare/@times-components/android-app@0.16.19...@times-components/android-app@0.16.20) (2019-08-21)
+
+**Note:** Version bump only for package @times-components/android-app
+
+
+
+
+
+## [0.16.19](https://github.com/newsuk/times-components/compare/@times-components/android-app@0.16.18...@times-components/android-app@0.16.19) (2019-08-21)
+
+**Note:** Version bump only for package @times-components/android-app
+
+
+
+
+
+## [0.16.18](https://github.com/newsuk/times-components/compare/@times-components/android-app@0.16.17...@times-components/android-app@0.16.18) (2019-08-21)
+
+**Note:** Version bump only for package @times-components/android-app
+
+
+
+
+
 ## [0.16.17](https://github.com/newsuk/times-components/compare/@times-components/android-app@0.16.16...@times-components/android-app@0.16.17) (2019-08-21)
 
 **Note:** Version bump only for package @times-components/android-app