+++ packages/depend/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.23](https://github.com/newsuk/times-components/compare/@times-components/depend@2.1.22...@times-components/depend@2.1.23) (2019-08-21)
+
+**Note:** Version bump only for package @times-components/depend
+
+
+
+
+
 ## [2.1.22](https://github.com/newsuk/times-components/compare/@times-components/depend@2.1.21...@times-components/depend@2.1.22) (2019-08-07)
 
 **Note:** Version bump only for package @times-components/depend